requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
pillow
-e .